# Conflicts:
#	trading-gustos/src/main/java/ru/gustos/trading/GustosBotLogic3.kt

+++ trading-bot/src/main/kotlin/ru/efreet/trading/logic/AbstractBotLogicParams.kt
@@ -4,8 +4,8 @@ package ru.efreet.trading.logic
  * Created by fluder on 20/02/2018.
  */
 interface AbstractBotLogicParams {
-    var stopLoss: Double
-    var tStopLoss: Double
-    var takeProfit: Double
-    var tTakeProfit: Double
+    var stopLoss: Float
+    var tStopLoss: Float
+    var takeProfit: Float
+    var tTakeProfit: Float
 }
