+++ trading-common/src/main/kotlin/ru/efreet/trading/bot/TradesStats.kt
@@ -3,20 +3,20 @@ package ru.efreet.trading.bot
 import java.time.ZonedDateTime
 
 data class TradesStats(var trades: Int = 0,
-                       var goodTrades: Double = 0.0,
-                       var profit: Double = 0.0,
-                       var profitPerTrade: Double = 0.0,
-                       var sdProfitPerTrade: Double = 0.0,
-                       var sma5: Double = 0.0,
-                       var sma10: Double = 0.0,
-                       var pearson: Double = 0.0,
+                       var goodTrades: Float = 0.0F,
+                       var profit: Float = 0.0F,
+                       var profitPerTrade: Float = 0.0F,
+                       var sdProfitPerTrade: Float = 0.0F,
+                       var sma5: Float = 0.0F,
+                       var sma10: Float = 0.0F,
+                       var pearson: Float = 0.0F,
                        var start: ZonedDateTime,
                        var end: ZonedDateTime,
-                       var profitPerDay: Double = 0.0,
-                       var relProfit: Double = 0.0
+                       var profitPerDay: Float = 0.0F,
+                       var relProfit: Float = 0.0F
 )
 
 data class TradesStatsShort(var trades: Int = 0,
-                       var profit: Double = 0.0,
-                       var pearson: Double = 0.0
+                       var profit: Float = 0.0F,
+                       var pearson: Float = 0.0F
 )
