+++ trading-common/src/main/kotlin/ru/efreet/trading/ta/indicators/XCachedIndicator.kt
@@ -5,25 +5,25 @@ package ru.efreet.trading.ta.indicators
  */
 abstract class XCachedIndicator<B>(bars: List<B>, val prop: BarGetterSetter<B>) : XAbstractIndicator<B>(bars) {
 
-    abstract fun calculate(index: Int, bar: B): Double
+    abstract fun calculate(index: Int, bar: B): Float
 
     open fun prepare() {
         for (i in 0 until bars.size)
             getValue(i)
     }
 
-    protected fun setPropValue(bar: B, value: Double) {
+    protected fun setPropValue(bar: B, value: Float) {
         prop.set(bar, value)
     }
 
-    protected fun getPropValue(bar: B): Double {
+    protected fun getPropValue(bar: B): Float {
         return prop.get(bar)
     }
 
-    override fun getValue(index: Int): Double {
+    override fun getValue(index: Int): Float {
         val bar = bars[index]
         var v = getPropValue(bar)
-        if (v == Double.MAX_VALUE) {
+        if (v == Float.MAX_VALUE) {
             v = calculate(index, bar)
             setPropValue(bar, v)
         }
