+++ trading-common/src/main/kotlin/ru/efreet/trading/trainer/MetricaTerm.kt
@@ -1,3 +1,3 @@
 package ru.efreet.trading.trainer
 
-data class MetricaTerm(val name: String, val value: Double)
+data class MetricaTerm(val name: String, val value: Float)
